ChatTalkActivity + Update code for design popup keyboard
