+++ app/src/main/java/jp/bap/traning/simplechat/ui/FriendAdapter.java
@@ -48,7 +48,7 @@ public class FriendAdapter extends RecyclerView.Adapter {
     public void onBindViewHolder(RecyclerView.ViewHolder holder, int position) {
         User user = mListUser.get(position);
         FriendViewHolder friendholder = (FriendViewHolder) holder;
-        friendholder.mUserName.setText(user.getFirstName());
+        friendholder.mUserName.setText(user.getFirstName()+" "+user.getLastName());
         friendholder.mAvatar.setOnClickListener(view -> mListener.onChat(user));
         friendholder.mImgButtonCall.setOnClickListener(view -> mListener.onCallAudio(user.getId()));
         friendholder.mImgButtonCallVideo.setOnClickListener(view -> mListener.onCallVideo(user.getId()));
