+++ app/src/main/res/layout/activity_chat_talks.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:id="@+id/relativeLayoutChatTalks"
     tools:context=".ui.ChatTalksActivity">
 
     <jp.bap.traning.simplechat.widget.CustomToolbar_
